Allow knowledgeable users to disable validating trove-classifiers
Classifiers are based on a "blessed list" of search terms that are
allowed on https://pypi.org and need to be regularly kept up to date in
order to validate them.

Many people don't care about this. Arguably, *no one* cares about this,
since wheels that have search terms that PyPI doesn't consider popular
enough will simply fail uploading to PyPI. But also, not everyone wants
to download new lists of "allowed words" from the internet every time
they check to see if e.g. pyproject.toml contains a valid format that
won't traceback when someone tries to read the "name" field and gets an
integer instead of a string. Or their entrypoints are malformed because
they aren't a valid python object reference.

This is also an issue because one might have an old version of the
classifiers cached, and then a new classifier is added to
https://pypi.org and you want to use it immediately, and the local
validator in the form of validate_pyproject fails but actually uploading
a wheel to https://pypi.org would work fine.